add back TestConfiguration dependency to IntegrationTestConfiguration; move Users class back to auth directory
