+++ scripts/run
@@ -54,6 +54,8 @@ RUNTYPE=${1}
 
 run_tests() {
   cd "${ROOT_DIR}"
+  eval "$("${SCRIPTS_DIR}"/render-configs.sh -e)"
+
   # shellcheck disable=SC2086 # The array syntax and quoting seems to break gradle
   ./gradlew testUnit ${GRADLE_ARGS}
 
@@ -83,6 +85,8 @@ run_connected_tests() {
   TEST_ENV=${TEST_ENV:-local}
 
   cd "${ROOT_DIR}"
+  eval "$("${SCRIPTS_DIR}"/render-configs.sh -e)"
+
   # shellcheck disable=SC2086 # The array syntax and quoting seems to break gradle
   ./gradlew testConnected ${GRADLE_ARGS}
